requirements.txt
python-telegram-bot[job-queue]==20.7
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
